Set command for week agenda & scheduled TODOs
The default custom command bound to "n" displays all TODOs, but I'm
rarely interested in TODOs which are SCHEDULED in the future.  This
setting will display a TODO list with items that are either SCHEDULED
for today, or don't have the SCHEDULED property.  Note, week agenda will
still display future SCHEDULED items, so this keeps the best of both
worlds.